requirements/requirements.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

numpy
pandas
scikit-learn

torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
lightning==2.3.3
speechbrain

kaldiio
soundfile

hydra-core==1.3.2
